Inline several of the functions for handling types. Also remove kind and type_id attributes from ndt.type.
