+++ dynd/src/type_functions.cpp
@@ -9,182 +9,9 @@
 #include "ctypes_interop.hpp"
 #include "utility_functions.hpp"
 
-#include <dynd/types/convert_type.hpp>
-#include <dynd/types/fixed_string_type.hpp>
-#include <dynd/types/string_type.hpp>
-#include <dynd/types/bytes_type.hpp>
-#include <dynd/types/pointer_type.hpp>
-#include <dynd/types/struct_type.hpp>
-#include <dynd/types/fixed_dim_type.hpp>
-#include <dynd/types/date_type.hpp>
-#include <dynd/types/time_type.hpp>
-#include <dynd/types/datetime_type.hpp>
-#include <dynd/types/type_type.hpp>
-#include <dynd/shape_tools.hpp>
-
-// Python's datetime C API
-#include "datetime.h"
-
 using namespace std;
 using namespace dynd;
 using namespace pydynd;
-
-void pydynd::init_type_functions()
-{
-  // Initialize the pydatetime API
-  PyDateTime_IMPORT;
-}
-
-inline void print_generic_type_repr(ostream &o, const ndt::type &d)
-{
-  stringstream sstmp;
-  sstmp << d;
-  o << "ndt.type(";
-  print_escaped_utf8_string(o, sstmp.str(), false);
-  o << ")";
-}
-
-PyObject *pydynd::_type_get_shape(const dynd::ndt::type &d)
-{
-  size_t ndim = d.get_ndim();
-  if (ndim > 0) {
-    dimvector shape(ndim);
-    d.extended()->get_shape(ndim, 0, shape.get(), NULL, NULL);
-    return intptr_array_as_tuple(ndim, shape.get());
-  }
-  else {
-    return PyTuple_New(0);
-  }
-}
-
-PyObject *pydynd::_type_get_kind(const dynd::ndt::type &d)
-{
-  stringstream ss;
-  ss << d.get_kind();
-  std::string s = ss.str();
-#if PY_VERSION_HEX >= 0x03000000
-  return PyUnicode_FromStringAndSize(s.data(), s.size());
-#else
-  return PyString_FromStringAndSize(s.data(), s.size());
-#endif
-}
-
-PyObject *pydynd::_type_get_type_id(const dynd::ndt::type &d)
-{
-  stringstream ss;
-  ss << d.get_type_id();
-  std::string s = ss.str();
-#if PY_VERSION_HEX >= 0x03000000
-  return PyUnicode_FromStringAndSize(s.data(), s.size());
-#else
-  return PyString_FromStringAndSize(s.data(), s.size());
-#endif
-}
-
-/**
- * Creates a dynd type out of typical Python typeobjects.
- */
-static dynd::ndt::type make__type_from_pytypeobject(PyTypeObject *obj)
-{
-  if (obj == &PyBool_Type) {
-    return ndt::make_type<bool1>();
-#if PY_VERSION_HEX < 0x03000000
-  }
-  else if (obj == &PyInt_Type) {
-    return ndt::make_type<int32_t>();
-#endif
-  }
-  else if (obj == &PyLong_Type) {
-    return ndt::make_type<int32_t>();
-  }
-  else if (obj == &PyFloat_Type) {
-    return ndt::make_type<double>();
-  }
-  else if (obj == &PyComplex_Type) {
-    return ndt::make_type<dynd::complex<double>>();
-  }
-  else if (obj == &PyUnicode_Type) {
-    return ndt::make_type<dynd::ndt::string_type>();
-  }
-  else if (obj == &PyByteArray_Type) {
-    return ndt::bytes_type::make(1);
-#if PY_VERSION_HEX >= 0x03000000
-  }
-  else if (obj == &PyBytes_Type) {
-    return ndt::bytes_type::make(1);
-#else
-  }
-  else if (obj == &PyString_Type) {
-    return ndt::make_type<dynd::ndt::string_type>();
-#endif
-  }
-  else if (PyObject_IsSubclass((PyObject *)obj, ctypes.PyCData_Type)) {
-    // CTypes type object
-    return _type_from_ctypes_cdatatype((PyObject *)obj);
-  }
-  else if (obj == PyDateTimeAPI->DateType) {
-    return ndt::date_type::make();
-  }
-  else if (obj == PyDateTimeAPI->TimeType) {
-    return ndt::time_type::make();
-  }
-  else if (obj == PyDateTimeAPI->DateTimeType) {
-    return ndt::datetime_type::make();
-  }
-
-  stringstream ss;
-  ss << "could not convert the Python TypeObject ";
-  pyobject_ownref obj_repr(PyObject_Repr((PyObject *)obj));
-  ss << pystring_as_string(obj_repr.get());
-  ss << " into a dynd type";
-  throw dynd::type_error(ss.str());
-}
-
-dynd::ndt::type pydynd::make__type_from_pyobject(PyObject *obj)
-{
-  if (DyND_PyType_Check(obj)) {
-    return ((DyND_PyTypeObject *)obj)->v;
-#if PY_VERSION_HEX < 0x03000000
-  }
-  else if (PyString_Check(obj)) {
-    return ndt::type(pystring_as_string(obj));
-  }
-  else if (PyInt_Check(obj)) {
-    return ndt::type(static_cast<type_id_t>(PyInt_AS_LONG(obj)));
-#endif
-  }
-  else if (PyLong_Check(obj)) {
-    return ndt::type(static_cast<type_id_t>(PyLong_AsLong(obj)));
-  }
-  else if (PyUnicode_Check(obj)) {
-    return ndt::type(pystring_as_string(obj));
-  }
-  else if (DyND_PyArray_Check(obj)) {
-    return ((DyND_PyArrayObject *)obj)->v.as<ndt::type>();
-  }
-  else if (PyType_Check(obj)) {
-#if DYND_NUMPY_INTEROP
-    ndt::type result;
-    if (_type_from_numpy_scalar_typeobject((PyTypeObject *)obj, result) == 0) {
-      return result;
-    }
-#endif // DYND_NUMPY_INTEROP
-    return make__type_from_pytypeobject((PyTypeObject *)obj);
-  }
-
-#if DYND_NUMPY_INTEROP
-  if (PyArray_DescrCheck(obj)) {
-    return _type_from_numpy_dtype((PyArray_Descr *)obj);
-  }
-#endif // DYND_NUMPY_INTEROP
-
-  stringstream ss;
-  ss << "could not convert the object ";
-  pyobject_ownref repr(PyObject_Repr(obj));
-  ss << pystring_as_string(repr.get());
-  ss << " into a dynd type";
-  throw dynd::type_error(ss.str());
-}
 
 static string_encoding_t encoding_from_pyobject(PyObject *encoding_obj)
 {
