+++ dynd/tests/test_dtype.py
@@ -234,23 +234,23 @@ class TestDType(unittest.TestCase):
         tp = ndt.type('var * 3 * 2 * int32')
         self.assertEqual(tp.shape, (-1, 3, 2))
 
-    def test_symbolic_type(self):
-        tp = ndt.type('(int, real) -> complex')
-        self.assertEqual(tp.type_id, 'callable')
-        self.assertEqual(nd.as_py(tp.pos_types), [ndt.int32, ndt.float64])
+    #def test_symbolic_type(self):
+    #    tp = ndt.type('(int, real) -> complex')
+    #    self.assertEqual(tp.type_id, 'callable')
+    #    self.assertEqual(nd.as_py(tp.pos_types), [ndt.int32, ndt.float64])
 #        self.assertEqual(tp.return_type, ndt.complex_float64)
-        tp = ndt.type('MyType')
+    #    tp = ndt.type('MyType')
   #      self.assertEqual(tp.type_id, 'typevar')
  #       self.assertEqual(tp.name, 'MyType')
-        tp = ndt.type('MyDim * int')
-        self.assertEqual(tp.type_id, 'typevar_dim')
+    #    tp = ndt.type('MyDim * int')
+    #    self.assertEqual(tp.type_id, 'typevar_dim')
    #     self.assertEqual(tp.name, 'MyDim')
     #    self.assertEqual(tp.element_type, ndt.int32)
-        tp = ndt.type('... * int')
-        self.assertEqual(tp.type_id, 'ellipsis_dim')
+    #    tp = ndt.type('... * int')
+    #    self.assertEqual(tp.type_id, 'ellipsis_dim')
      #   self.assertEqual(tp.element_type, ndt.int32)
-        tp = ndt.type('MyEll... * int')
-        self.assertEqual(tp.type_id, 'ellipsis_dim')
+    #    tp = ndt.type('MyEll... * int')
+    #    self.assertEqual(tp.type_id, 'ellipsis_dim')
        # self.assertEqual(tp.name, 'MyEll')
       #  self.assertEqual(tp.element_type, ndt.int32)
